Move Sierra types from rpc pgk to contracts
