+++ contracts/contracts_test.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/NethermindEth/juno/core/felt"
 	internalUtils "github.com/NethermindEth/starknet.go/internal/utils"
-	"github.com/NethermindEth/starknet.go/rpc"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -47,7 +46,7 @@ func TestUnmarshalContractClass(t *testing.T) {
 		content, err := os.ReadFile(testCase.CasmPath)
 		require.NoError(t, err)
 
-		var class rpc.ContractClass
+		var class ContractClass
 		err = json.Unmarshal(content, &class)
 		require.NoError(t, err)
 		assert.Equal(t, class.SierraProgram[0].String(), testCase.SierraProgram0)
@@ -228,5 +227,53 @@ func newNestedFieldArray(val ...NestedUints) *NestedUints {
 		IsArray: true,
 		Value:   nil,
 		Values:  val,
+	}
+}
+
+// TestDeprecatedContractClass_UnmarshalValidJSON_Successful is a test function that
+// tests the successful unmarshalling of valid JSON into a DeprecatedContractClass
+// object.
+//
+// It reads the content of a file, then unmarshals the content into a
+// DeprecatedContractClass object using the json.Unmarshal function. If any error
+// occurs during the process, the test fails.
+//
+// Parameters:
+// - t: The testing.T object used for reporting test failures and logging.
+// Returns:
+//
+//	none
+func TestDeprecatedContractClass_UnmarshalValidJSON_Successful(t *testing.T) {
+	content, err := os.ReadFile("./tests/0x01b661756bf7d16210fc611626e1af4569baa1781ffc964bd018f4585ae241c1.json")
+	if err != nil {
+		t.Fatal("should be able to read file", err)
+	}
+
+	contractClass := DeprecatedContractClass{}
+	if err := json.Unmarshal(content, &contractClass); err != nil {
+		t.Fatal("should be able unmarshall Class", err)
+	}
+}
+
+// TestContractClass_UnmarshalValidJSON_Successful is a test function that validates the successful unmarshalling of valid JSON data into a ContractClass object.
+//
+// The function does the following:
+// - Reads the content of a file specified by the validContractCompiledPath variable.
+// - Unmarshals the content into a ContractClass object using the json.Unmarshal function.
+//
+// Parameters:
+// - t: The testing.T object used for reporting test failures and logging.
+// Returns:
+//
+//	none
+func TestContractClass_UnmarshalValidJSON_Successful(t *testing.T) {
+	content, err := os.ReadFile("./tests/0x03e9b96873987da76121f74a3df71e38c44527d8ce2ad115bcfda3cba0548cc3.json")
+	if err != nil {
+		t.Fatal("should be able to read file", err)
+	}
+
+	contractClass := ContractClass{}
+	if err := json.Unmarshal(content, &contractClass); err != nil {
+		t.Fatal("should be able unmarshall Class", err)
 	}
 }
